chore: separated project creation
