+++ src/app/(app)/globals.css
@@ -80,6 +80,25 @@
   filter: blur(1.5rem);
   opacity: 0.5;
 }
+.empty-project-card:after,
+.empty-project-card:before {
+  content: '';
+  position: absolute;
+  top: -2px;
+  left: -2px;
+  right: -2px;
+  bottom: -2px;
+  background-image: conic-gradient(from var(--angle), transparent 75%, #7c3aed);
+  z-index: -1;
+  border-radius: 0.375rem;
+  animation: spin 8s linear infinite;
+}
+
+.empty-project-card::before {
+  filter: blur(1.5rem);
+  opacity: 0.5;
+}
+
 @keyframes spin {
   from {
     --angle: 0deg;
