+++ src/components/dashboard/List.tsx
@@ -5,24 +5,32 @@ import { Skeleton } from '../common/Skeleton'
 import { Project } from '@payload-types'
 import Link from 'next/link'
 
+import EmptyProjects from './EmptyProjects'
+
 // const iconList = Object.keys(Icons)
 const List = ({
   projects,
   isLoading,
+  isProjectsEmpty,
 }: {
   projects: Project[]
   isLoading: boolean
+  isProjectsEmpty: boolean
 }) => {
   return (
-    <div className='z-10 grid grid-cols-1 items-center gap-4 md:grid-cols-2 lg:grid-cols-3'>
+    <div>
       {isLoading ? (
-        <>
+        <div className='z-10 grid grid-cols-1 items-center gap-4 md:grid-cols-2 lg:grid-cols-3'>
           <Skeleton className='h-48 w-full' />
           <Skeleton className='h-48 w-full' />
           <Skeleton className='h-48 w-full' />
-        </>
+        </div>
+      ) : isProjectsEmpty ? (
+        <EmptyProjects />
       ) : (
-        projects?.map((site, index) => <ListItem key={index} site={site} />)
+        <div className='z-10 grid grid-cols-1 items-center gap-4 md:grid-cols-2 lg:grid-cols-3'>
+          {projects?.map((site, index) => <ListItem key={index} site={site} />)}
+        </div>
       )}
     </div>
   )
