+++ src/components/dashboard/validator.ts
@@ -3,8 +3,8 @@ import { z } from 'zod'
 export const ProjectSchema = z.object({
   name: z
     .string()
-    .min(3, { message: 'Service name must be at least 3 characters long' })
-    .max(30, { message: 'Service name cannot exceed 30 characters' }),
+    .min(3, { message: 'Project name must be at least 3 characters long' })
+    .max(30, { message: 'Project name cannot exceed 30 characters' }),
   description: z
     .string()
     .min(6, { message: 'Project Description must be at least 6 characters' }),
